remove old oauth file
